Add temporary fix for bug in craft ui when scroll didn't reset when other crafting category shown

+++ src/ui/craft_ui.hpp
@@ -41,7 +41,7 @@ namespace Terrarium {
         ItemCellRenderer item_cell_renderer;
         sf::ConvexShape arrow;
 
-        unsigned int current_scroll = 0;
+        size_t current_scroll = 0;
 
     public:
         CraftUI(const Gfx &gfx);
